change BpClickable to a functional component
